Remove error emittors
