+++ api/web/src/components/Home.vue
@@ -13,26 +13,18 @@
         </div>
     </div>
 
-    <TablerError v-if='err' :err='err' @close='err = null'/>
     <PageFooter/>
 </div>
 </template>
 
 <script>
-import { TablerError } from '@tak-ps/vue-tabler';
 import PageFooter from './PageFooter.vue';
 import CardConnections from './cards/Connections.vue';
 import CardLayers from './cards/Layers.vue';
 
 export default {
     name: 'Home',
-    data: function() {
-        return {
-            err: false,
-        }
-    },
     components: {
-        TablerError,
         PageFooter,
         CardConnections,
         CardLayers
